@@ -1,26 +1,8 @@
 {
   "name": "cryptocurrency-tracker",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "private": true,
   "homepage": "https://alexjcm.github.io/cryptocurrency-tracker",
-  "dependencies": {
-    "@emotion/core": "^10.3.1",
-    "@kurkle/color": "^0.3.2",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.1",
-    "@testing-library/user-event": "^12.2.0",
-    "axios": "^1.3.4",
-    "chart.js": "^3.9.1",
-    "chartjs-adapter-date-fns": "^2.0.1",
-    "date-fns": "^2.29.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.7.1",
-    "react-pure-loaders": "^3.0.1",
-    "react-router-dom": "^5.3.4",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^2.1.0"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -29,6 +11,26 @@
     "docker-run": "docker run --rm -p 3000:3000 --name cryptocurrency-tracker-app cryptocurrency-tracker",
     "docker-stop": "docker rm -f cryptocurrency-tracker-app",
     "docker-build-prod": "docker build --rm -t alexjcm/cryptocurrency-tracker ."
+  },
+  "dependencies": {
+    "@emotion/core": "^10.3.1",
+    "@kurkle/color": "^0.3.2",
+    "axios": "^1.3.4",
+    "chart.js": "^3.9.1",
+    "chartjs-adapter-date-fns": "^2.0.1",
+    "date-fns": "^2.29.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
+    "react-pure-loaders": "^3.0.1",
+    "react-router-dom": "^5.3.4",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^3.3.0"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^14.4.3"
   },
   "eslintConfig": {
     "extends": [
